+++ README.md
@@ -1,3 +1,5 @@
+Docs after api is somewhat stabilized..
+
 Rapid Development..
 
 Retry task, control concurrency etc..
