+++ packages/retry/jest.json
@@ -1,6 +1,6 @@
 {
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": "../",
+  "rootDir": "./",
   "testEnvironment": "node",
   "testRegex": ".spec.ts$",
   "transform": {
