+++ packages/retry/package.json
@@ -6,7 +6,7 @@
   "types": "./dist/index.d.ts",
   "devDependencies": {},
   "scripts": {
-    "test": "jest --config ./test/jest.json",
+    "test": "jest --config ./jest.json",
     "lint": "eslint \"{src,test}/**/*.ts\""
   }
 }
