+++ packages/retry/src/utils/delay.ts
@@ -1,0 +1,5 @@
+export const delay = (props: { delay: number; currentAttempt: number; maxAttempts: number }) => {
+  return new Promise((resolve) => {
+    setTimeout(() => resolve(props.delay), props.delay);
+  });
+};
