+++ packages/retry/src/utils/isRetryable.ts
@@ -1,0 +1,3 @@
+export const isRetryable = (currentAttempt: number, maxAttempts: number) => {
+  return currentAttempt < maxAttempts;
+};
